!Split coupling code to splited modules
